Move scheduler interface under builtins (#139)
* Move JobScheduler interface to builtins

* Fixing examples for the job

(cherry picked from commit 7f879048442cd97697ee22b8e858d325578d6606)
