+++ calm/dsl/builtins/__init__.py
@@ -93,7 +93,7 @@ from .models.endpoint import Endpoint, _endpoint, CalmEndpoint
 
 from .models.vm_profile import VmProfile
 from .models.vm_blueprint import VmBlueprint
-from .models.job import Job
+from .models.job import Job, JobScheduler
 
 __all__ = [
     "Ref",
@@ -194,6 +194,7 @@ __all__ = [
     "_endpoint",
     "CalmEndpoint",
     "AppProtection",
+    "JobScheduler",
     "AhvVmRecoveryResources",
     "ahv_vm_recovery_spec",
     "Job",
