+++ tests/jobs/job_one_time_app_action.py
@@ -1,5 +1,4 @@
-from calm.dsl.builtins import Job
-from calm.dsl.cli.scheduler import JobScheduler
+from calm.dsl.builtins import Job, JobScheduler
 
 start_date_time = "2021-11-18 17:58:15"
 time_zone = "Asia/Calcutta"
