+++ tests/scheduler/expiry_less_currentdate_recurring.py
@@ -1,7 +1,6 @@
 import uuid
 
-from calm.dsl.builtins import Job
-from calm.dsl.cli.scheduler import JobScheduler
+from calm.dsl.builtins import Job, JobScheduler
 from datetime import datetime, timedelta
 
 start_date_time = "2050-10-08 16:17:15"
