+++ tests/scheduler/job_list.py
@@ -1,5 +1,4 @@
-from calm.dsl.builtins import Job
-from calm.dsl.cli.scheduler import JobScheduler
+from calm.dsl.builtins import Job, JobScheduler
 
 start_date_time = "2050-10-08 16:17:15"
 time_zone = "America/Jamaica"
